moved PodInstanceStatus generator into raml conversions package
