+++ src/main/scala/mesosphere/marathon/raml/RamlConversions.scala
@@ -1,6 +1,6 @@
 package mesosphere.marathon.raml
 
 trait RamlConversions extends ConstraintConversion with ContainerConversion
-  with EnvVarConversion with NetworkConversion with PodConversion
+  with EnvVarConversion with NetworkConversion with PodConversion with PodStatusConversion
 
 object RamlConversions extends RamlConversions
